bf16 (except 3D force fp32)

+++ examples/opensora_hpcai/tools/caption/PLLaVA/models/pllava/configuration_pllava.py
@@ -23,7 +23,7 @@ class PllavaConfig:
         use_pooling=True,
         gradient_checkpointing=False,
         pad_token_id=0,
-        dtype=ms.float32,
+        dtype=ms.bfloat16,
         **kwargs,
     ):
         self.ignore_index = ignore_index
